Move plugin activator to a separate class file
